requirements.txt
Django==3.2.15
djangorestframework==3.12.4
django-filter==22.1
django-cors-headers==3.13.0
django-redis==5.2.0
mysqlclient==2.1.1
redis==4.3.4
requests==2.31.0
urllib3==1.26.12
pyjwt==2.4.0
PyYAML==6.0
Markdown==3.4.1